+++ yorkshire4/profiles/__init__.py
@@ -1,0 +1,4 @@
+
+from .commodore64 import COMMODORE_64
+from .microbee import MICROBEE
+from .zxspectrum import ZXSPECTRUM
